Moved test to new package
